Move PCPCheckAndHeal into its own file
